+++ src/tracing/index.js
@@ -27,6 +27,7 @@ exports.init = function(_config) {
       require('./instrumentation/mongodb.js').init(config);
       require('./instrumentation/kafka.js').init(config);
       require('./instrumentation/mysql.js').init(config);
+      require('./instrumentation/redis.js').init(config);
     }
   }
 };
@@ -85,6 +86,7 @@ exports.activate = function() {
       require('./instrumentation/mongodb.js').activate();
       require('./instrumentation/kafka.js').activate();
       require('./instrumentation/mysql.js').activate();
+      require('./instrumentation/redis.js').activate();
     }
   }
 };
@@ -99,6 +101,7 @@ exports.deactivate = function() {
       require('./instrumentation/elasticsearch.js').deactivate();
       require('./instrumentation/httpServer.js').deactivate();
       require('./instrumentation/httpClient.js').deactivate();
+      require('./instrumentation/redis.js').deactivate();
     }
 
     require('./opentracing').deactivate();