+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+ - Add redis instrumentation.
+
 ## 1.30.3
  - HTTP tracing will report incorrect path in case of express.js router usage.